refactor(frontend): move onboarding tour state to general state module
