+++ apps/frontend/app/lib/state/general.ts
